+++ paylogic/lookups.py
@@ -15,7 +15,7 @@ class TargetBranchesView(Select2View):
     def get_results(self, request, term, page, context):
         """Get the list of possible target branch(es)."""
         case_id = self.kwargs['case_id']
-        _, case_title, original_branch, feature_branch, ci_project = views.get_fogbugz_case_info(case_id)
+        _, case_title, original_branch, feature_branch, ci_project = views.get_fogbugz_case_info(request, case_id)
 
         target_vcs, target_url, target_revision, target_branch_is_local, _ = views.parse_branch_vcs_info(
             original_branch, settings.ORIGINAL_BRANCH_DEFAULT_PREFIX)